+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `v1.8.1` 
+
+- **changed** some internal projection operations to use the new `projection` function in SimpleSDMLayers [#557]
+- **removed** an internal function that was essentially `boundingbox`
+
 ## `v1.8.0`
 
 - **changed** a lot of the presentation of the documentation